# Use lazy iter for expr.UnexpandedSeriesSet