--- a/promql/engine.go
+++ b/promql/engine.go
@@ -818,6 +818,37 @@ func extractGroupsFromPath(p []parser.Node) (bool, []string) {
 	return false, nil
 }
 
+type concreteSeriesSet struct {
+	cur               int
+	expandedSeriesSet []storage.Series
+}
+
+func (c *concreteSeriesSet) Next() bool {
+	c.cur++
+	return c.cur-1 < len(c.expandedSeriesSet)
+}
+
+func (c *concreteSeriesSet) At() storage.Series {
+	return c.expandedSeriesSet[c.cur-1]
+}
+
+func (c *concreteSeriesSet) Err() error {
+	return nil
+}
+
+func (c *concreteSeriesSet) Warnings() storage.Warnings {
+	return nil
+}
+
+func lazilyExpandSeriesSet(expr *parser.VectorSelector) storage.SeriesSet {
+	if expr.Series != nil {
+		return &concreteSeriesSet{
+			expandedSeriesSet: expr.Series,
+		}
+	}
+	return expr.UnexpandedSeriesSet
+}
+
 func checkAndExpandSeriesSet(ctx context.Context, expr parser.Expr) (storage.Warnings, error) {
 	switch e := expr.(type) {
 	case *parser.MatrixSelector:
@@ -1232,6 +1263,7 @@ func (ev *evaluator) eval(expr parser.Expr) (parser.Value, storage.Warnings) {
 		var (
 			matrixArgIndex int
 			matrixArg      bool
+			hadSubquery    bool
 			warnings       storage.Warnings
 		)
 		for i := range e.Args {
@@ -1240,6 +1272,7 @@ func (ev *evaluator) eval(expr parser.Expr) (parser.Value, storage.Warnings) {
 			if _, ok := a.(*parser.MatrixSelector); ok {
 				matrixArgIndex = i
 				matrixArg = true
+				hadSubquery = true
 				break
 			}
 			// parser.SubqueryExpr can be used in place of parser.MatrixSelector.
@@ -1282,11 +1315,20 @@ func (ev *evaluator) eval(expr parser.Expr) (parser.Value, storage.Warnings) {
 		sel := unwrapStepInvariantExpr(e.Args[matrixArgIndex]).(*parser.MatrixSelector)
 		selVS := sel.VectorSelector.(*parser.VectorSelector)
 
-		ws, err := checkAndExpandSeriesSet(ev.ctx, sel)
-		warnings = append(warnings, ws...)
-		if err != nil {
-			ev.error(errWithWarnings{errors.Wrap(err, "expanding series"), warnings})
+		if hadSubquery {
+			ws, err := checkAndExpandSeriesSet(ev.ctx, sel)
+			warnings = append(warnings, ws...)
+			if err != nil {
+				ev.error(errWithWarnings{errors.Wrap(err, "expanding series"), warnings})
+			}
 		}
+
+		seriesSet := lazilyExpandSeriesSet(selVS)
+		if err := seriesSet.Err(); err != nil {
+			ev.error(errWithWarnings{errors.Wrap(err, "expanding series"), seriesSet.Warnings()})
+		}
+
+		warnings = append(warnings, seriesSet.Warnings()...)
 		mat := make(Matrix, 0, len(selVS.Series)) // Output matrix.
 		offset := durationMilliseconds(selVS.Offset)
 		selRange := durationMilliseconds(sel.Range)
@@ -1301,11 +1343,12 @@ func (ev *evaluator) eval(expr parser.Expr) (parser.Value, storage.Warnings) {
 		enh := &EvalNodeHelper{Out: make(Vector, 0, 1)}
 		// Process all the calls for one time series at a time.
 		it := storage.NewBuffer(selRange)
-		for i, s := range selVS.Series {
+		for seriesSet.Next() {
+			s := seriesSet.At()
 			ev.currentSamples -= len(points)
 			points = points[:0]
 			it.Reset(s.Iterator())
-			metric := selVS.Series[i].Labels()
+			metric := s.Labels()
 			// The last_over_time function acts like offset; thus, it
 			// should keep the metric name.  For all the other range
 			// vector functions, the only change needed is to drop the
@@ -1317,7 +1360,7 @@ func (ev *evaluator) eval(expr parser.Expr) (parser.Value, storage.Warnings) {
 				Metric: metric,
 				Points: getPointSlice(numSteps),
 			}
-			inMatrix[0].Metric = selVS.Series[i].Labels()
+			inMatrix[0].Metric = s.Labels()
 			for ts, step := ev.startTimestamp, -1; ts <= ev.endTimestamp; ts += ev.interval {
 				step++
 				// Set the non-matrix arguments.
@@ -1471,16 +1514,19 @@ func (ev *evaluator) eval(expr parser.Expr) (parser.Value, storage.Warnings) {
 		return String{V: e.Val, T: ev.startTimestamp}, nil
 
 	case *parser.VectorSelector:
-		ws, err := checkAndExpandSeriesSet(ev.ctx, e)
-		if err != nil {
-			ev.error(errWithWarnings{errors.Wrap(err, "expanding series"), ws})
+		seriesSet := lazilyExpandSeriesSet(e)
+		if err := seriesSet.Err(); err != nil {
+			ev.error(errWithWarnings{errors.Wrap(err, "expanding series"), seriesSet.Warnings()})
 		}
+
 		mat := make(Matrix, 0, len(e.Series))
 		it := storage.NewMemoizedEmptyIterator(durationMilliseconds(ev.lookbackDelta))
-		for i, s := range e.Series {
+
+		for seriesSet.Next() {
+			s := seriesSet.At()
 			it.Reset(s.Iterator())
 			ss := Series{
-				Metric: e.Series[i].Labels(),
+				Metric: s.Labels(),
 				Points: getPointSlice(numSteps),
 			}
 
@@ -1502,7 +1548,7 @@ func (ev *evaluator) eval(expr parser.Expr) (parser.Value, storage.Warnings) {
 				putPointSlice(ss.Points)
 			}
 		}
-		return mat, ws
+		return mat, seriesSet.Warnings()
 
 	case *parser.MatrixSelector:
 		if ev.startTimestamp != ev.endTimestamp {
@@ -1606,15 +1652,16 @@ func (ev *evaluator) vectorSelector(node *parser.VectorSelector, ts int64) (Vect
 	if err != nil {
 		ev.error(errWithWarnings{errors.Wrap(err, "expanding series"), ws})
 	}
+
 	vec := make(Vector, 0, len(node.Series))
 	it := storage.NewMemoizedEmptyIterator(durationMilliseconds(ev.lookbackDelta))
-	for i, s := range node.Series {
+	for _, s := range node.Series {
 		it.Reset(s.Iterator())
 
 		t, v, ok := ev.vectorSelectorSingle(it, node, ts)
 		if ok {
 			vec = append(vec, Sample{
-				Metric: node.Series[i].Labels(),
+				Metric: s.Labels(),
 				Point:  Point{V: v, T: t},
 			})
 
@@ -1630,6 +1677,7 @@ func (ev *evaluator) vectorSelector(node *parser.VectorSelector, ts int64) (Vect
 
 // vectorSelectorSingle evaluates a instant vector for the iterator of one time series.
 func (ev *evaluator) vectorSelectorSingle(it *storage.MemoizedSeriesIterator, node *parser.VectorSelector, ts int64) (int64, float64, bool) {
+
 	refTime := ts - durationMilliseconds(node.Offset)
 	var t int64
 	var v float64
@@ -1684,19 +1732,19 @@ func (ev *evaluator) matrixSelector(node *parser.MatrixSelector) (Matrix, storag
 
 		it = storage.NewBuffer(durationMilliseconds(node.Range))
 	)
-	ws, err := checkAndExpandSeriesSet(ev.ctx, node)
-	if err != nil {
-		ev.error(errWithWarnings{errors.Wrap(err, "expanding series"), ws})
+	seriesSet := lazilyExpandSeriesSet(vs)
+	if err := seriesSet.Err(); err != nil {
+		ev.error(errWithWarnings{errors.Wrap(err, "expanding series"), seriesSet.Warnings()})
 	}
 
-	series := vs.Series
-	for i, s := range series {
+	for seriesSet.Next() {
+		s := seriesSet.At()
 		if err := contextDone(ev.ctx, "expression evaluation"); err != nil {
 			ev.error(err)
 		}
 		it.Reset(s.Iterator())
 		ss := Series{
-			Metric: series[i].Labels(),
+			Metric: s.Labels(),
 		}
 
 		ss.Points = ev.matrixIterSlice(it, mint, maxt, getPointSlice(16))
@@ -1707,7 +1755,7 @@ func (ev *evaluator) matrixSelector(node *parser.MatrixSelector) (Matrix, storag
 			putPointSlice(ss.Points)
 		}
 	}
-	return matrix, ws
+	return matrix, seriesSet.Warnings()
 }
 
 // matrixIterSlice populates a matrix vector covering the requested range for a