@darshanime 2021-08-14
is it possible to run memory benchmarks on this pr? ideally it should improve memory metrics...