@darshanime 2021-08-18
It looks like we are trading allocs and memory for compute. Not sure if it is significant though...
<details>
  <summary>see benchmark</summary>

```
benchmark                                                                                                              old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_one,steps=1-12                                                                              12704          12587          -0.92%
BenchmarkRangeQuery/expr=a_one,steps=10-12                                                                             13021          12987          -0.26%
BenchmarkRangeQuery/expr=a_one,steps=100-12                                                                            19171          18991          -0.94%
BenchmarkRangeQuery/expr=a_one,steps=1000-12                                                                           62463          60888          -2.52%
BenchmarkRangeQuery/expr=a_ten,steps=1-12                                                                              49616          49153          -0.93%
BenchmarkRangeQuery/expr=a_ten,steps=10-12                                                                             53631          53216          -0.77%
BenchmarkRangeQuery/expr=a_ten,steps=100-12                                                                            114326         113358         -0.85%
BenchmarkRangeQuery/expr=a_ten,steps=1000-12                                                                           545443         525098         -3.73%
BenchmarkRangeQuery/expr=a_hundred,steps=1-12                                                                          434592         423143         -2.63%
BenchmarkRangeQuery/expr=a_hundred,steps=10-12                                                                         478598         462108         -3.45%
BenchmarkRangeQuery/expr=a_hundred,steps=100-12                                                                        1090305        1056330        -3.12%
BenchmarkRangeQuery/expr=a_hundred,steps=1000-12                                                                       5496274        5195543        -5.47%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-12                                                                    18068          18283          +1.19%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-12                                                                   19137          19290          +0.80%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-12                                                                  33605          33539          -0.20%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-12                                                                 157137         147929         -5.86%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-12                                                                    56857          55580          -2.25%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-12                                                                   68130          67321          -1.19%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-12                                                                  213560         205367         -3.84%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-12                                                                 1403833        1334073        -4.97%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-12                                                                459552         463309         +0.82%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-12                                                               572371         566998         -0.94%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-12                                                              2029293        1952052        -3.81%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-12                                                             13868705       13345258       -3.77%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10000-12                                                                1540640        1488155        -3.41%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-12                                                                14958439       14676736       -1.88%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10000-12                                                            151777400      145966047      -3.83%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-12                                                  806545         806423         -0.02%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-12                                                 1295965        1302655        +0.52%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-12                                                6254928        6250383        -0.07%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-12                                               56269292       56200015       -0.12%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-12                                                  6464568        6328313        -2.11%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-12                                                 11276743       11474245       +1.75%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-12                                                60984989       61371332       +0.63%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-12                                               557415860      558309378      +0.16%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-12                                              65380587       62882171       -3.82%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-12                                             112909449      113174193      +0.23%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-12                                            610954906      608567882      -0.39%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-12                                           5542577564     5538404261     -0.08%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-12                                                                 643803         654538         +1.67%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-12                                                                915671         797566         -12.90%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-12                                                               2498531        2445520        -2.12%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-12                                                              19609463       19067768       -2.76%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-12                                                                 5740460        5517708        -3.88%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-12                                                                7396645        7198328        -2.68%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-12                                                               24400004       24113290       -1.18%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-12                                                              192950511      189781505      -1.64%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-12                                                             57291867       55589936       -2.97%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-12                                                            73852258       71714096       -2.90%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-12                                                           243189185      236591359      -2.71%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-12                                                          1910706505     1890818900     -1.04%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-12                                                                    625309         622957         -0.38%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-12                                                                   715714         704612         -1.55%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-12                                                                  1641693        1571036        -4.30%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-12                                                                 11176322       10264036       -8.16%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-12                                                                    5437992        5404702        -0.61%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-12                                                                   6461561        6214864        -3.82%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-12                                                                  15825787       14982210       -5.33%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-12                                                                 108009151      102962417      -4.67%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-12                                                                55504928       54208445       -2.34%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-12                                                               64318404       64173804       -0.22%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-12                                                              159843585      150441446      -5.88%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-12                                                             1072876001     1010090315     -5.85%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1-12                                                        606458         601145         -0.88%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=10-12                                                       622340         618679         -0.59%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=100-12                                                      808958         812005         +0.38%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-12                                                     2645318        2723961        +2.97%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-12                                                        5320582        5208468        -2.11%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-12                                                       5524321        5381409        -2.59%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-12                                                      7445187        7384968        -0.81%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-12                                                     26718749       26582328       -0.51%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-12                                                    53353319       52130253       -2.29%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-12                                                   55281063       54107117       -2.12%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-12                                                  74721698       73615231       -1.48%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-12                                                 265387054      264494065      -0.34%
BenchmarkRangeQuery/expr=-a_one,steps=1-12                                                                             14460          14261          -1.38%
BenchmarkRangeQuery/expr=-a_one,steps=10-12                                                                            14909          14645          -1.77%
BenchmarkRangeQuery/expr=-a_one,steps=100-12                                                                           21147          20804          -1.62%
BenchmarkRangeQuery/expr=-a_one,steps=1000-12                                                                          65500          63059          -3.73%
BenchmarkRangeQuery/expr=-a_ten,steps=1-12                                                                             55370          54790          -1.05%
BenchmarkRangeQuery/expr=-a_ten,steps=10-12                                                                            59730          59901          +0.29%
BenchmarkRangeQuery/expr=-a_ten,steps=100-12                                                                           121629         120015         -1.33%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-12                                                                          569389         540633         -5.05%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-12                                                                         479263         459499         -4.12%
BenchmarkRangeQuery/expr=-a_hundred,steps=10-12                                                                        516504         500020         -3.19%
BenchmarkRangeQuery/expr=-a_hundred,steps=100-12                                                                       1142105        1115557        -2.32%
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-12                                                                      5641435        5310523        -5.87%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-12                                                                      24957          24929          -0.11%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-12                                                                     31287          30961          -1.04%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-12                                                                    98324          95535          -2.84%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-12                                                                   726364         706155         -2.78%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-12                                                                      123761         121479         -1.84%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-12                                                                     192972         186445         -3.38%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-12                                                                    899847         866491         -3.71%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-12                                                                   7634627        7176588        -6.00%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-12                                                              1149080        1100834        -4.20%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-12                                                             1851724        1774159        -4.19%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-12                                                            9563303        9085954        -4.99%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-12                                                           83405809       78811204       -5.51%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-12                                                                  7382092        7193070        -2.56%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-12                                                                  78175354       74714641       -4.43%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-12                                                          876323217      843055658      -3.80%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-12                                                     57033          56819          -0.38%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-12                                                    59027          59452          +0.72%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-12                                                   89972          90766          +0.88%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-12                                                  375712         368893         -1.81%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-12                                                     131407         130201         -0.92%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-12                                                    160411         159230         -0.74%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-12                                                   477939         477455         -0.10%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-12                                                  3367139        3363300        -0.11%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-12                                             825260         812061         -1.60%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-12                                            1165295        1144768        -1.76%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-12                                           4799674        4792211        -0.16%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-12                                          39288690       39281162       -0.02%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-12                                                      58642          56712          -3.29%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-12                                                     61417          60805          -1.00%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-12                                                    106612         106467         -0.14%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-12                                                   507854         502920         -0.97%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1-12                                                      133946         131822         -1.59%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=10-12                                                     174769         173549         -0.70%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=100-12                                                    606828         605865         -0.16%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1000-12                                                   4592168        4716084        +2.70%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-12                                              862245         858390         -0.45%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-12                                             1315907        1284960        -2.35%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-12                                            5912648        5952051        +0.67%
…